fix in nhapkho module & finish xuat kho module

+++ database/migrations/2021_05_17_024731_create_mat_hangs_table.php
@@ -16,7 +16,7 @@ class CreateMatHangsTable extends Migration
         Schema::create('mat_hangs', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('tenMH');
-            $table->float('don_gia');
+            $table->bigInteger('don_gia');
             $table->string('nhaCC')->nullable(true);
             $table->timestamps();
         });
